Make GeoFileTileSource base class
